[refactor #37] 폴더 구조 변경 및 adoc 파일 생성
