+++ src/main/java/com/project/dogwalker/common/config/WebConfig.java
@@ -2,8 +2,8 @@ package com.project.dogwalker.common.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.project.dogwalker.common.resolver.queryString.QueryStringArgsResolver;
-import com.project.dogwalker.member.controller.AuthArgumentResolver;
-import com.project.dogwalker.member.controller.AuthInterceptor;
+import com.project.dogwalker.common.resolver.auth.AuthArgumentResolver;
+import com.project.dogwalker.common.interceptor.AuthInterceptor;
 import com.project.dogwalker.member.token.JwtTokenProvider;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
